backend/requirements.txt
gradient-adk
langgraph
langchain-core
gradient
playwright
httpx
google-genai
fastapi
uvicorn
